src/user/requirements.txt
backoff==2.2.1
build==0.10.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Deprecated==1.2.13
docutils==0.19
Flask==2.2.2
fluent-logger==0.10.0
googleapis-common-protos==1.56.2
grpcio==1.51.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lorem-text==2.1
MarkupSafe==2.1.2
morecontext==0.6.0
msgpack==1.0.4
opentelemetry-api==1.16.0
opentelemetry-distro==0.37b0
opentelemetry-exporter-otlp==1.16.0
opentelemetry-exporter-otlp-proto-grpc==1.16.0
opentelemetry-exporter-otlp-proto-http==1.16.0
opentelemetry-instrumentation==0.37b0
opentelemetry-instrumentation-aws-lambda==0.37b0
opentelemetry-instrumentation-dbapi==0.37b0
opentelemetry-instrumentation-flask==0.37b0
opentelemetry-instrumentation-jinja2==0.37b0
opentelemetry-instrumentation-logging==0.37b0
opentelemetry-instrumentation-requests==0.37b0
opentelemetry-instrumentation-sqlite3==0.37b0
opentelemetry-instrumentation-urllib==0.37b0
opentelemetry-instrumentation-wsgi==0.37b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.16.0
opentelemetry-sdk==1.16.0
opentelemetry-semantic-conventions==0.37b0
opentelemetry-util-http==0.37b0
packaging==23.0
process-tilt-ciphersmaug @ https://github.com/ciphersmaug/process_tilt_logger/releases/download/alpha/process_tilt_ciphersmaug-0.0.1-py3-none-any.whl
protobuf==3.20.3
pyproject_hooks==1.0.0
requests==2.28.2
six==1.16.0
thrift==0.16.0
tilt==0.0.5
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
waitress==2.1.2
Werkzeug==2.2.2
wrapt==1.15.0
